Move projects, tasks and comment related functions for backend API v2
