+++ server/api/messaging/message_apis.py
@@ -11,74 +11,6 @@ from server.services.messaging.message_service import (
 )
 from server.services.users.authentication_service import token_auth, tm
 
-
-class ProjectsMessageAll(Resource):
-    @tm.pm_only()
-    @token_auth.login_required
-    def post(self, project_id):
-        """
-        Send message to all contributors to a project
-        ---
-        tags:
-            - messaging
-        produces:
-            - application/json
-        parameters:
-            - in: header
-              name: Authorization
-              description: Base64 encoded session token
-              required: true
-              type: string
-              default: Token sessionTokenHere==
-            - name: project_id
-              in: path
-              description: The unique project ID
-              required: true
-              type: integer
-              default: 1
-            - in: body
-              name: body
-              required: true
-              description: JSON object for creating draft project
-              schema:
-                  properties:
-                      subject:
-                          type: string
-                          default: Thanks
-                          required: true
-                      message:
-                          type: string
-                          default: Thanks for your contribution
-                          required: true
-        responses:
-            200:
-                description: All mapped tasks validated
-            401:
-                description: Unauthorized - Invalid credentials
-            500:
-                description: Internal Server Error
-        """
-        try:
-            message_dto = MessageDTO(request.get_json())
-            message_dto.from_user_id = tm.authenticated_user_id
-            message_dto.validate()
-        except DataError as e:
-            current_app.logger.error(f"Error validating request: {str(e)}")
-            return str(e), 400
-
-        try:
-            threading.Thread(
-                target=MessageService.send_message_to_all_contributors,
-                args=(project_id, message_dto),
-            ).start()
-
-            return {"Success": "Messages started"}, 200
-        except Exception as e:
-            error_msg = f"Send message all - unhandled error: {str(e)}"
-            current_app.logger.critical(error_msg)
-            return {"error": error_msg}, 500
-
-
 class HasNewMessages(Resource):
     @tm.pm_only(False)
     @token_auth.login_required
