+++ server/api/projects/contributions.py
@@ -1,0 +1,38 @@
+from flask_restful import Resource, current_app, request
+from server.services.stats_service import StatsService, NotFound
+from server.services.project_service import ProjectService
+from server.services.users.user_service import UserService
+
+
+class ProjectsContributionsAPI(Resource):
+    def get(self, project_id):
+        """
+        Get all user contributions on a project
+        ---
+        tags:
+          - projects
+        produces:
+          - application/json
+        parameters:
+            - name: project_id
+              in: path
+              required: true
+              type: integer
+              default: 1
+        responses:
+            200:
+                description: User contributions
+            404:
+                description: No contributions
+            500:
+                description: Internal Server Error
+        """
+        try:
+            contributions = StatsService.get_user_contributions(project_id)
+            return contributions.to_primitive(), 200
+        except NotFound:
+            return {"Error": "No contributions on project"}, 404
+        except Exception as e:
+            error_msg = f"User GET - unhandled error: {str(e)}"
+            current_app.logger.critical(error_msg)
+            return {"error": error_msg}, 500
