+++ server/api/users/user_apis.py
@@ -357,52 +357,6 @@ class UserOSMAPI(Resource):
             return {"error": error_msg}, 500
 
 
-class UserMappedProjects(Resource):
-    def get(self, username):
-        """
-        Gets projects user has mapped
-        ---
-        tags:
-          - user
-        produces:
-          - application/json
-        parameters:
-            - in: header
-              name: Accept-Language
-              description: Language user is requesting
-              type: string
-              required: true
-              default: en
-            - name: username
-              in: path
-              description: The users username
-              required: true
-              type: string
-              default: Thinkwhere
-        responses:
-            200:
-                description: Mapped projects found
-            404:
-                description: No mapped projects found
-            500:
-                description: Internal Server Error
-        """
-        try:
-            locale = (
-                request.environ.get("HTTP_ACCEPT_LANGUAGE")
-                if request.environ.get("HTTP_ACCEPT_LANGUAGE")
-                else "en"
-            )
-            user_dto = UserService.get_mapped_projects(username, locale)
-            return user_dto.to_primitive(), 200
-        except NotFound:
-            return {"Error": "User or mapping not found"}, 404
-        except Exception as e:
-            error_msg = f"User GET - unhandled error: {str(e)}"
-            current_app.logger.critical(error_msg)
-            return {"error": error_msg}, 500
-
-
 class UserSetRole(Resource):
     @tm.pm_only()
     @token_auth.login_required
